Minor refactoring and cleanup
